persisting filters now
